requirements.txt
# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu118
# python -m pip install paddlepaddle-gpu==2.6.0 -i https://pypi.tuna.tsinghua.edu.cn/simple
numpy==1.26.4
utils==1.0.2
imutils==0.5.4
pillow==11.0.0
deep-sort-realtime==1.3.2
# paddlepaddle==2.6.2
paddleocr==2.9.1
ultralytics==8.3.37
Flask==3.0.2
Flask-Cors==5.0.0
grpcio-tools==1.46.0